requirements.txt
click==8.1.3
joblib==1.3.1
lxml==4.9.3
mlflow
numpy==1.24.4
pandas==1.5.3
Pillow==10.0.0
python-dotenv==1.0.0
rich
scikit_learn
scipy==1.11.1
SQLAlchemy==1.4.41
typing_extensions
psycopg2-binary
streamlit
nbformat
seaborn
plotly
kaleido
category_encoders
zenml
zenml[server]
